[FE] feat: TextArea 상태 분리
